requirements.txt
boto3==1.14.58
botocore==1.17.58
certifi==2020.6.20
chardet==3.0.4
docutils==0.15.2
idna==2.10
jmespath==0.10.0
plotly==4.9.0
python-dateutil==2.8.1
python-decouple==3.3
requests==2.24.0
retrying==1.3.3
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
